2020 extrusion STL files.
These are the stl's for the 2020 extrusions.
listed in the name is the length of the extrusion, and the number that need to be printed. 

the biggest printer that i have at the time of this upload is 300mm. So i will need to split the extrusion in to sizes that will fit on the printer. Then they will be glued together.

Not sure what glue will be used at this time.